crowdstrike-falcon-azure-sentinel-data-connector/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
boto3
requests
filesplit
